List Columns

Returns a list of columns for the given table/view in the specified database. The result includes the name, description, dataType, whether it is nullable or if it is partitioned and if it is broken in buckets.

list_columns(sc, table, database = NULL)

Arguments

sc

A spark_connection.

table

character(1). The name of the table.

database

character(1). The name of the database for which the functions should be listed (default: NULL).

Value

A tibble with 6 columns:

name - The name of the column.

description - Description of the column.

dataType - The column data type.

nullable - Whether the column is nullable or not.

isPartition - Whether the column is partitioned or not.

isBucket - Whether the column is broken in buckets.

Examples

if (FALSE) {
sc <- sparklyr::spark_connect(master = "local")
mtcars_spark <- sparklyr::copy_to(dest = sc, df = mtcars)
list_columns(sc = sc, table = "mtcars")
}